DRY site contained entities concern. Ensure site and institution attributes are related. Complete site_prefix for proper descendant search. Refactor device and test_result models. Add site_prefix to encounter model.
