+++ app/models/test_result.rb
@@ -2,6 +2,7 @@ class TestResult < ActiveRecord::Base
   include AutoUUID
   include Entity
   include Resource
+  include SiteContained
 
   NAME_FIELD = 'name'
   LAB_USER_FIELD = 'site_user'
@@ -12,8 +13,6 @@ class TestResult < ActiveRecord::Base
   has_many :test_result_parsed_data
 
   belongs_to :device
-  belongs_to :institution
-  belongs_to :site
   belongs_to :sample_identifier, inverse_of: :test_results, autosave: true
   belongs_to :patient
   belongs_to :encounter
@@ -25,7 +24,7 @@ class TestResult < ActiveRecord::Base
   validate :validate_encounter
   validate :validate_patient
 
-  before_save   :set_foreign_keys
+  before_save   :set_foreign_keys, prepend: true
   before_save   :set_entity_id
   after_destroy :destroy_from_index
 
@@ -99,7 +98,6 @@ class TestResult < ActiveRecord::Base
   def set_foreign_keys
     self.site_id = device.try(:site_id)
     self.institution_id = device.try(:institution_id)
-    self.site_prefix = device.try(:site).try(:prefix)
   end
 
   def set_entity_id
