+++ db/migrate/20160112190503_add_site_prefix_to_encounters.rb
@@ -1,0 +1,5 @@
+class AddSitePrefixToEncounters < ActiveRecord::Migration
+  def change
+    add_column :encounters, :site_prefix, :string
+  end
+end
